my_flask/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
rococo
bcrypt
PyJWT
Flask-CORS
mailjet_rest